refactor/cleanup test_available
